% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czekanowski_index.R
\name{czekanowski_index}
\alias{czekanowski_index}
\title{Czekanowski index for overlap in resource use between two species.}
\usage{
czekanowski_index(p, q)
}
\arguments{
\item{p}{A vector of relative frequence of activity pattern.}

\item{q}{A vector of relative frequence of activity pattern.}
}
\value{
An index value of overlap resource use.
}
\description{
Czekanowski index for overlap in resource use between two species.
}
\examples{
n = 6
p = ifelse(runif(n) > 0.5, 1, 0)
p <- p/sum(p)
q = ifelse(runif(n) > 0.5, 1, 0)
q <- q/sum(q)
czekanowski_index(p, q)
}
