% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1.R
\docType{data}
\name{ex1}
\alias{ex1}
\title{rosario example 1}
\format{
## `ex1`
A matrix with 12 temporal bins (columns) and 5 species:
}
\source{
<https://>
}
\usage{
ex1
}
\description{
A matrix with temporal ocupancy of 5 species.
}
\keyword{datasets}
