% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_null_model.R
\name{get_null_model}
\alias{get_null_model}
\title{Niche overlap null model. Resampling through rosario algorithm
to preserve the temporal autocorrelation.}
\usage{
get_null_model(mat, method, nsim = 100, parallel = FALSE)
}
\arguments{
\item{mat}{A matrix with species per time columns.}

\item{method}{A measure of temporal niche overlap between pairs of species}

\item{nsim}{The number of simulations}

\item{parallel}{Logical. Do the resampling on parallel}
}
\value{
A matrix with pairwise overlaping temporal resource use.
}
\description{
Niche overlap null model. Resampling through rosario algorithm
to preserve the temporal autocorrelation.
}
\examples{
temp_overlap(ex1, method = "pianka")

get_null_model(ex1, method = "pianka", nsim = 10)
}
