% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pianka_index.R
\name{pianka_index}
\alias{pianka_index}
\title{Pianka's index}
\usage{
pianka_index(p, q)
}
\arguments{
\item{p}{A vector of relative frequence of activity pattern.}

\item{q}{A vector of relative frequence of activity pattern.}
}
\value{
An index value of overlap resource use.
}
\description{
Pianka's index
}
\examples{
n = 100
p = ifelse(runif(n) > 0.5, 1, 0)
q = ifelse(runif(n) > 0.5, 1, 0)
pianka_index(p, q)
}
