% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_matrix.R
\name{rescale_matrix}
\alias{rescale_matrix}
\title{Rescale matrix
This function  rescale the rows of a matrix dividing each element of a row over the sum of the row.}
\usage{
rescale_matrix(m)
}
\arguments{
\item{m}{A numeric matrix to rescale}
}
\value{
A rescaled matrix
}
\description{
Rescale matrix
This function  rescale the rows of a matrix dividing each element of a row over the sum of the row.
}
\examples{
ex1_rescale <- rescale_matrix(ex1)
}
