% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_overlap.R
\name{temp_overlap}
\alias{temp_overlap}
\title{Assemblage-wide temporal niche segregation}
\usage{
temp_overlap(mat, method = c("pianka", "czekanowski"))
}
\arguments{
\item{mat}{A matrix with species per time columns.}

\item{method}{A measure of temporal niche overlap between pairs of species}
}
\value{
The mean temporal overlaping index from a pairwise distance matrix
}
\description{
Assemblage-wide temporal niche segregation
}
\examples{
temp_overlap(ex1, method = "pianka")
ex1_rescale <- rescale_matrix(ex1)
temp_overlap(ex1_rescale, method = "czekanowski")
}
