% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_overlap_df.R
\name{temp_overlap_df}
\alias{temp_overlap_df}
\title{Convert a square tempral overlapping matrix to a pairwise distance dataframe}
\usage{
temp_overlap_df(mat)
}
\arguments{
\item{mat}{A pairwise  square matrix of temporal niche overlapping of species}
}
\value{
A data.frame if the distance between species
}
\description{
Convert a square tempral overlapping matrix to a pairwise distance dataframe
}
\examples{
d <- temp_overlap_matrix(ex1)
temp_overlap_df(d)
}
