% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_overlap_matrix.R
\name{temp_overlap_matrix}
\alias{temp_overlap_matrix}
\title{Assemblage-wide temporal niche segregation}
\usage{
temp_overlap_matrix(mat, method = c("pianka", "czekanowski"))
}
\arguments{
\item{mat}{A matrix with species per time columns.}

\item{method}{A measure of temporal niche overlap between pairs of species}
}
\value{
A matrix with pairwise overlaping temporal resource use.
}
\description{
Assemblage-wide temporal niche segregation
}
\examples{
temp_overlap_matrix(ex1, method = "pianka")
ex1_rescale <- rescale_matrix(ex1)
temp_overlap_matrix(ex1_rescale, method = "czekanowski")
}
