% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_permutation.R
\name{vec_permutation}
\alias{vec_permutation}
\title{Arbitrary ciclic permutation of a numeric vector}
\usage{
vec_permutation(numvec, x = 1)
}
\arguments{
\item{numvec}{A numeric vector. This function is designed to deal with
the randomization
time series of animal activity}

\item{x}{The new starting position of the permutation}
}
\value{
A permuted vector
}
\description{
Arbitrary ciclic permutation of a numeric vector
}
\examples{
vec_permutation(1:24, 4)
}
