% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosario.R
\name{rosario}
\alias{rosario}
\title{rosario}
\usage{
rosario(mat)
}
\arguments{
\item{numvec}{A numeric vector with activity pattern time series}
}
\value{
A list with all the permutations in the time series
}
\description{
rosario
}
\examples{
rosario(c(40, 25, 18, 10, 5, 2))
}
