% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosario.R
\name{rosario_sample}
\alias{rosario_sample}
\title{Random sampling a niche overlap matrix preserving the
autocorrelation temporal structure.
Each row of the matrix is permuted starting in a random
column. The result also could be reversed with a
binomial probability of 0.5.}
\usage{
rosario_sample(mat)
}
\arguments{
\item{mat}{A matrix with species (rows) and temporal intervals (columns)
to be sampled preserving the temporal autocorrelation of the intervals}
}
\value{
A random matrix with same number of species and temporal intervals
}
\description{
Random sampling a niche overlap matrix preserving the
autocorrelation temporal structure.
Each row of the matrix is permuted starting in a random
column. The result also could be reversed with a
binomial probability of 0.5.
}
\examples{
rosario_sample(ex1)
}
